requirements.txt
requests==2.22.0
urllib3==1.25.3
PyInstaller==3.4
Pillow==6.0.0